requirements.txt
asgiref==3.2.7
autopep8==1.5.2
boto3==1.12.46
botocore==1.15.46
dj-database-url==0.5.0
Django==3.0.5
django-crispy-forms==1.9.1
django-storages==1.9.1
django-tinymce==3.0.2
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.5
mysqlclient==1.4.6
Pillow==7.1.2
pycodestyle==2.5.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
